backend/requirements.txt
networkx
fn == 0.4.3
cadCAD
pandas
numpy
matplotlib
hvplot
Django==4.1
djangorestframework>=3.12,<4
django-cors-headers>=3.10,<4
psycopg2-binary>=2.9,<3
gunicorn
dj-database-url
whitenoise
